requirements.txt
# requirements.txt
pyspark
requests
psycopg2-binary